java: raise an exception on conversion of unknown polymorphic types in Java SDK
Rather than silently succeeding (returning None), raise an exception.

This brings the Java SDK in line with the behaviour of the Go SDK.

Includes some rudimentary test coverage to demonstrate this works for one of those types.

JAVADEPS was updated solely for the change of io.grpc from a runtime to an
implementation dependency for the tests, which was required to construct a GRPC
exception to exercise the exception conversion code.

GitOrigin-RevId: 6743b0e76616eaee0e9fe1f8583bd4f123092a5b